% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_boxplots.R
\name{get_boxplots}
\alias{get_boxplots}
\title{get_boxplots}
\usage{
get_boxplots(
  long_matrix,
  my_x,
  my_y,
  my_fill,
  my_y_positions,
  my_x_min,
  my_x_max,
  my_annotation,
  my_palette,
  log_y = FALSE,
  cell_type = TRUE
)
}
\arguments{
\item{long_matrix}{The long (melted) version of a cell adundance matrix, (e.g.:)
as obtained from CIBERSORT or MCP-counter}

\item{my_y_positions}{for plotting p-values. My_y_positions vectors as
obtained from the function get_pvalues_positions}

\item{my_x_min}{for plotting p-values. my_x_min vectors as
obtained from the function get_pvalues_positions}

\item{my_x_max}{for plotting p-values. my_x_max vectors as
obtained from the function get_pvalues_positions}

\item{my_annotation}{for plotting p-values. my_annotation vectors as
obtained from the function get_pvalues_positions}

\item{my_palette}{A vector of color to be passed to the ggplot functions. If
provided by the user, it must be the same length of number of factors in
sample_groups.}

\item{log_y}{y-axis to be plotted as log10. Useful to plot absolute counts (
i.e.: as from MCP-counter).
Default: FALSE}

\item{cell_type}{if TRUE, it expects a long_matrix with cell_types and it
generates a faceted plot, once facet for each cell-type. Stylistic choice
to make the cell type names readable}

\item{x}{feature in the x axis. Used as well to facet wrap if cell_type == TRUE}

\item{y}{feature in the y axis}

\item{fill}{feature to be used for the fill}
}
\description{
get_boxplots
}
\details{
This function generates boxplots for the main figure.
}
\note{
wrapper for p-values positions to be annotated

2024-11-22
}
\author{
Andrea Del Cortona \href{mailto:andrea.delcortona@gmail.com}{andrea.delcortona@gmail.com}
}
