% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvalues_positions.R
\name{get_pvalues_positions}
\alias{get_pvalues_positions}
\title{get_pvalues_positions}
\usage{
get_pvalues_positions(long_matrix, ttest)
}
\arguments{
\item{long_matrix}{The long (melted) version of a cell adundance matrix, (e.g.:)
as obtained from CIBERSORT or MCP-counter}

\item{ttest}{A t-test table obtained from rstatix::t_test}
}
\description{
get_pvalues_positions
}
\details{
I get the x and y positions and the adj-p-value significance to be annotated
on ggplot boxplots for different cell types across the groups.
I replace similar functions in ggpubr and ggsign packages that seems not to
work well for me.
}
\note{
wrapper for p-values positions to be annotated

2024-11-15
}
\author{
Andrea Del Cortona \href{mailto:andrea.delcortona@gmail.com}{andrea.delcortona@gmail.com}
}
