% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infiltR-package.R, R/infiltR.R
\docType{package}
\name{infiltR}
\alias{infiltR-package}
\alias{infiltR}
\title{infiltR}
\usage{
infiltR(
  counts_table,
  metadata,
  sample_groups,
  my_palette = "default",
  save_plots = TRUE,
  outdir = "default",
  mcp_featuresType = "HUGO_symbols",
  mcp_probesets = "default",
  mcp_genes = "default",
  cb_perm = 500,
  cb_pval_thr = 0.1,
  cb_QN = FALSE,
  qs_is_arraydata = FALSE,
  qs_is_tumordata = TRUE,
  qs_scale_mRNA = TRUE,
  qs_method = "lsei",
  qs_rm_genes = "default"
)
}
\arguments{
\item{counts_table}{A RNAseq counts table, with genes as rows and samples as
columns. The count table is expected to be already preprocessed (i.e.: lowly
abundant genes removed with \code{edgeR::filterByExpr()} and \code{cpm()}
normalized).}

\item{metadata}{A metadata matrix with samples as rows}

\item{sample_groups}{The columns of the metadata matrix to be used for the
comparisons between sample groups. E.g.: a column indicating which sample
was treated and which one not.}

\item{my_palette}{A vector of color to be passed to the ggplot functions. If
provided by the user, it must be the same length of number of factors in
sample_groups.
Default: "default", it uses standard ggplot2 palette.}

\item{save_plots}{save plots in pdf and png format
Default: TRUE}

\item{outdir}{Output directory of the plots
Default: "default", it prints in the current working directory}

\item{mcp_featuresType}{MCP-counter parameter.
featuresType of the counts_table to be passed to MCP-counter functions.
Possible values: "affy133P2_probesets", "HUGO_symbols", "ENTREZ_ID", "ENSEMBL_ID".
Default: "HUGO_symbols"}

\item{mcp_probesets}{MCP-counter parameter.
The probe set that indicates the genes of the immune infiltrating cells
Default: "default", which retrieves the original MCP-counter repo with
\code{read.table(curl:::curl("https://raw.githubusercontent.com/ebecht/MCPcounter/master/Signatures/probesets.txt"),sep="\t",stringsAsFactors=FALSE,colClasses="character")}}

\item{mcp_genes}{MCP-counter parameter.
The gene set that indicates the genes of the immune infiltrating cells
Default: "default", which retrieves the original MCP-counter repo with
\code{read.table(curl:::curl("https://raw.githubusercontent.com/ebecht/MCPcounter/master/Signatures/genes.txt"),sep="\t",stringsAsFactors=FALSE,header=TRUE,colClasses="character",check.names=FALSE}}

\item{cb_perm}{CIBERSORT parameter.
Number of permutations to be performed to get a p-value on the estimated
infiltrating immune cells.
Default: 100}

\item{cb_pval_thr}{CIBERSORT parameter.
P-value threshold to retain samples for comparisons.
Default: 0.1}

\item{cb_QN}{CIBERSORT parameter.
Perform Quantile Normalization on the counts_table.
Default: FALSE}

\item{qs_is_arraydata}{quanTIseq parameter
If it is microarray or RNAseq data.
Default: FALSE}

\item{qs_scale_mRNA}{quanTIseq parameter
Logical value. If set to FALSE, it disables the correction of cell-type-specific
mRNA content bias.
Default: TRUE}

\item{qs_method}{quanTIseq parameter
Character string, defining the deconvolution method to be used: lsei for
constrained least squares regression, hampel, huber, or bisquare for robust
regression with Huber, Hampel, or Tukey bisquare estimators, respectively.
Default: "lsei".}

\item{qs_rm_genes}{quanTIseq parameter
Character vector, specifying which genes have to be excluded from the
deconvolution analysis. It can be provided as:
• a vector of gene symbols (contained in the expression_data)
• a single string among the choices of "none" (no genes are removed) and
"default" (a list of genes with noisy expression RNA-seq data is removed,
as explained in the quanTIseq paper).
Default: "default" for RNA-seq data, "none" for microarrays.}
}
\description{
infiltR, a tool for tumor microenvironmnet characterization
}
\details{
InfiltR characterize the tumor microenvironment by evaluating the amount of
infiltrating immune cells of RNA-seq samples of solid tumor biopsies.
It relies on three complentary tools:
\itemize{
\item \href{https://genomebiology.biomedcentral.com/articles/10.1186/s13059-016-1070-5}{MCP-counter};
\item \href{https://pmc.ncbi.nlm.nih.gov/articles/PMC5895181/}{CIBERSORT};
\item \href{https://genomemedicine.biomedcentral.com/articles/10.1186/s13073-019-0638-6}{quanTIseq}.
}

InfiltR provides absolute and relative estimates of infiltrating immune cells,
it returns abundance tables for each tool that can be used for further downstream
analyses, and prints publication ready figures.
}
\note{
wrapper for infiltR, a tool for tumor microenvironmnet characterization

2024-11-13
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/19ADC99/infiltR}
}

}
\author{
\strong{Maintainer}: Andrea Del Cortona \email{andrea.delcortona@gmail.com}


Andrea Del Cortona \href{mailto:andrea.delcortona@gmail.com}{andrea.delcortona@gmail.com}
}
