% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cb_qc.R
\name{plot_cb_qc}
\alias{plot_cb_qc}
\title{plot_cb_qc}
\usage{
plot_cb_qc(infiltr_out, save_plots = TRUE, outdir = "default")
}
\arguments{
\item{infiltr_out}{infiltr_out output object}

\item{save_plots}{save plots in pdf and png format
Default: TRUE}

\item{outdir}{Output directory of the plots
Default: "default", it prints in the current working directory}
}
\description{
plot_cb_qc
}
\details{
This function plots QC statistics for CIBERSORT.
}
\note{
print CIBERSORT QC

2024-12-09
}
\author{
Andrea Del Cortona \href{mailto:andrea.delcortona@gmail.com}{andrea.delcortona@gmail.com}
}
