% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comparisons.R
\name{plot_comparisons}
\alias{plot_comparisons}
\title{plot_comparisons}
\usage{
plot_comparisons(infiltr_out, save_plots = TRUE, outdir = "default")
}
\arguments{
\item{infiltr_out}{infiltr_out output object}

\item{save_plots}{save plots in pdf and png format
Default: TRUE}

\item{outdir}{Output directory of the plots
Default: "default", it prints in the current working directory}
}
\description{
plot_comparisons
}
\details{
This function plots comparisons of infiltrating immune cells estimated with
MCP-counter, CIBERSORT and aunTIseq
}
\note{
plot estimates comparisons

2024-12-09
}
\author{
Andrea Del Cortona \href{mailto:andrea.delcortona@gmail.com}{andrea.delcortona@gmail.com}
}
