% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_infiltR.R
\name{plot_infiltR}
\alias{plot_infiltR}
\title{plot_infiltR}
\usage{
plot_infiltR(
  infiltr_out,
  metadata,
  sample_groups,
  my_palette,
  save_plots = TRUE,
  outdir = "default"
)
}
\arguments{
\item{infiltr_out}{infiltr_out output object}

\item{metadata}{A metadata matrix with samples as rows}

\item{sample_groups}{The columns of the metadata matrix to be used for the
comparisons between sample groups. E.g.: a column indicating which sample
was treated and which one not.}

\item{my_palette}{A vector of color to be passed to the ggplot functions. If
provided by the user, it must be the same length of number of factors in
sample_groups.
Default: "default", it uses standard ggplot2 palette.}

\item{save_plots}{save plots in pdf and png format
Default: TRUE}

\item{outdir}{Output directory of the plots
Default: "default", it prints in the current working directory}
}
\description{
plot_infiltR
}
\details{
This function plots infiltR estimates.
}
\note{
wrapper for plotting infiltR summary figures

2024-11-13
}
\author{
Andrea Del Cortona \href{mailto:andrea.delcortona@gmail.com}{andrea.delcortona@gmail.com}
}
