% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cibersort.R
\name{run_cibersort}
\alias{run_cibersort}
\title{run_cibersort}
\usage{
run_cibersort(
  infiltr_out,
  counts_table,
  metadata,
  sample_groups,
  cb_perm,
  cb_pval_thr,
  cb_QN
)
}
\arguments{
\item{infiltr_out}{infiltR output object}

\item{counts_table}{A RNAseq counts table, with genes as rows and samples as
columns. The count table is expected to be already preprocessed (i.e.: lowly
abundant genes removed with \code{edgeR::filterByExpr()} and \code{cpm()}
normalized).}

\item{metadata}{A metadata matrix with samples as rows}

\item{sample_groups}{The columns of the metadata matrix to be used for the
comparisons between sample groups. E.g.: a column indicating which sample
was treated and which one not.}

\item{cb_perm}{CIBERSORT parameter.
Number of permutations to be performed to get a p-value on the estimated
infiltrating immune cells.
Default: 500}

\item{cb_pval_thr}{CIBERSORT parameter.
P-value threshold to retain samples for comparisons.
Default: 0.1}

\item{cb_QN}{CIBERSORT parameter.
Perform Quantile Normalization on the counts_table.
Default: FALSE}
}
\description{
run_cibersort
}
\details{
Retro-engineered from https://www.nature.com/articles/nmeth.3337
}
\note{
wrapper for CIBERSORT

2024-11-15
}
\author{
Andrea Del Cortona \href{mailto:andrea.delcortona@gmail.com}{andrea.delcortona@gmail.com}
}
