% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mcpcounter.R
\name{run_mcpcounter}
\alias{run_mcpcounter}
\title{run_mcpcounter}
\usage{
run_mcpcounter(
  infiltr_out,
  counts_table,
  metadata,
  sample_groups,
  mcp_featuresType,
  mcp_probesets,
  mcp_genes
)
}
\arguments{
\item{infiltr_out}{infiltR output object}

\item{counts_table}{A RNAseq counts table, with genes as rows and samples as
columns. The count table is expected to be already preprocessed (i.e.: lowly
abundant genes removed with \code{edgeR::filterByExpr()} and \code{cpm()}
normalized).}

\item{metadata}{A metadata matrix with samples as rows}

\item{sample_groups}{The columns of the metadata matrix to be used for the
comparisons between sample groups. E.g.: a column indicating which sample
was treated and which one not.}

\item{mcp_featuresType}{MCP-counter parameter.
featuresType of the counts_table to be passed to MCP-counter functions.
Possible values: "affy133P2_probesets", "HUGO_symbols", "ENTREZ_ID", "ENSEMBL_ID".
Default: "HUGO_symbols"}

\item{mcp_probesets}{MCP-counter parameter.
The probe set that indicates the genes of the immune infiltrating cells
Default: "default", which retrieves the original MCP-counter repo with
\code{read.table(curl:::curl("https://raw.githubusercontent.com/ebecht/MCPcounter/master/Signatures/probesets.txt"),sep="\t",stringsAsFactors=FALSE,colClasses="character")}}

\item{mcp_genes}{MCP-counter parameter.
The gene set that indicates the genes of the immune infiltrating cells
Default: "default", which retrieves the original MCP-counter repo with
\code{read.table(curl:::curl("https://raw.githubusercontent.com/ebecht/MCPcounter/master/Signatures/genes.txt"),sep="\t",stringsAsFactors=FALSE,header=TRUE,colClasses="character",check.names=FALSE}}
}
\description{
run_mcpcounter
}
\details{
This function runs MCP-counter.
}
\note{
wrapper for MCP-counter

2024-11-13
}
\author{
Andrea Del Cortona \href{mailto:andrea.delcortona@gmail.com}{andrea.delcortona@gmail.com}
}
