% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_quantiseq.R
\name{run_quantiseq}
\alias{run_quantiseq}
\title{run_quantiseq}
\usage{
run_quantiseq(
  infiltr_out,
  counts_table,
  metadata,
  sample_groups,
  qs_is_arraydata,
  qs_is_tumordata,
  qs_scale_mRNA,
  qs_method,
  qs_rm_genes
)
}
\arguments{
\item{infiltr_out}{infiltR output object}

\item{counts_table}{A RNAseq counts table, with genes as rows and samples as
columns. The count table is expected to be already preprocessed (i.e.: lowly
abundant genes removed with \code{edgeR::filterByExpr()} and \code{cpm()}
normalized).}

\item{metadata}{A metadata matrix with samples as rows}

\item{qs_is_arraydata}{quanTIseq parameter
If it is microarray or RNAseq data.
Default: FALSE}

\item{qs_scale_mRNA}{quanTIseq parameter
Logical value. If set to FALSE, it disables the correction of cell-type-specific
mRNA content bias.
Default: TRUE}

\item{qs_method}{quanTIseq parameter
Character string, defining the deconvolution method to be used: lsei for
constrained least squares regression, hampel, huber, or bisquare for robust
regression with Huber, Hampel, or Tukey bisquare estimators, respectively.
Default: "lsei".}

\item{qs_rm_genes}{quanTIseq parameter
Character vector, specifying which genes have to be excluded from the
deconvolution analysis. It can be provided as:
• a vector of gene symbols (contained in the expression_data)
• a single string among the choices of "none" (no genes are removed) and
"default" (a list of genes with noisy expression RNA-seq data is removed,
as explained in the quanTIseq paper).
Default: "default" for RNA-seq data, "none" for microarrays.}
}
\description{
run_quantiseq
}
\details{
This function runs quanTIseq.
}
\note{
wrapper for MCP-counter

2024-11-13
}
\author{
Andrea Del Cortona \href{mailto:andrea.delcortona@gmail.com}{andrea.delcortona@gmail.com}
}
